pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wandb-plot-utils"
version = "0.0.1"
description = "Utilities for plotting Weights & Biases experiments"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "Sicelukwanda Zwane", email = "sicelukwanda.zwane@gmail.com" },
]
dependencies = [
  "wandb",
  "pandas",
  "plotnine",
  "matplotlib",
  "numpy",
  "seaborn",
  "hydra-core",
]

[project.scripts]
wandb-plot = "wandb_plot_utils.cli:main"
